Modularize project api methods
