+++ ansible_events_ui/api/project.py
@@ -1,0 +1,87 @@
+import sqlalchemy as sa
+from fastapi import APIRouter, Depends
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from ansible_events_ui.db.dependency import get_db_session
+from ansible_events_ui.db.models import (
+    extra_vars,
+    inventories,
+    playbooks,
+    projects,
+    rulebooks,
+)
+from ansible_events_ui.project import clone_project, sync_project
+from ansible_events_ui.schemas import Project
+
+router = APIRouter()
+
+
+@router.get("/api/projects/")
+async def list_projects(db: AsyncSession = Depends(get_db_session)):
+    query = sa.select(projects)
+    result = await db.execute(query)
+    return result.all()
+
+
+@router.post("/api/project/")
+async def create_project(
+    p: Project, db: AsyncSession = Depends(get_db_session)
+):
+    found_hash, tempdir = await clone_project(p.url, p.git_hash)
+    p.git_hash = found_hash
+    query = sa.insert(projects).values(url=p.url, git_hash=p.git_hash)
+    result = await db.execute(query)
+    (project_id,) = result.inserted_primary_key
+    await sync_project(project_id, tempdir, db)
+    await db.commit()
+    return {**p.dict(), "id": project_id}
+
+
+@router.get("/api/project/{project_id}")
+async def read_project(
+    project_id: int, db: AsyncSession = Depends(get_db_session)
+):
+    # FIXME(cutwater): Return HTTP 404 if project doesn't exist
+    query = sa.select(projects).where(projects.c.id == project_id)
+    project = (await db.execute(query)).first()
+
+    response = dict(project)
+
+    response["rulesets"] = (
+        await db.execute(
+            sa.select(rulebooks.c.id, rulebooks.c.name)
+            .select_from(rulebooks)
+            .join(projects)
+            .where(projects.c.id == project_id)
+        )
+    ).all()
+
+    response["inventories"] = (
+        await db.execute(
+            sa.select(inventories.c.id, inventories.c.name)
+            .select_from(inventories)
+            .join(projects)
+            .where(projects.c.id == project_id)
+        )
+    ).all()
+
+    response["vars"] = (
+        await db.execute(
+            sa.select(extra_vars.c.id, extra_vars.c.name)
+            .select_from(extra_vars)
+            .join(projects)
+            .where(projects.c.id == project_id)
+        )
+    ).all()
+
+    response["playbooks"] = (
+        await db.execute(
+            sa.select(playbooks.c.id, playbooks.c.name)
+            .select_from(playbooks)
+            .join(projects)
+            .where(projects.c.id == project_id)
+        )
+    ).all()
+
+    return response
+
